aomdec: Fix scalable streams in non-annex-b, .obu format
There is some special-case code in obudec.c to split each temporal unit
up into multiple "chunks", one per spatial layer. This is used by
examples/scalable_decoder.c to extract both spatial layers of its input,
so that further processing can be applied.

Unfortunately, in its current form, this code causes aomdec to break
when given a scalable stream. But it's fairly straightforward to adjust
the special-case code so that it isn't triggered by aomdec. Then we will
get the same behaviour as we would get if the stream were in annex-b format,
or in a container.

That is, aomdec will pass a full temporal unit at a time into libaom.
The current behaviour of libaom in this case is to output a single
frame from each temporal unit - the frame chosen is the last (and so
highest-spatial-layer) show or show-existing frame within the temporal
unit.

Given that the intent of spatial scalability is for each temporal
unit to contain multiple versions of the "same" picture, with different
sizes/qualities, this seems like the behaviour a "typical" application
would want, and is useful for conformance testing.

BUG=aomedia:1851

Change-Id: I61ddf62220aac849d86e5a621bdb516934dedd54