Fix a segmentation fault in multi-thread code
The segmentation fault happened while using arch=x86-linux-gcc, and
-DSANITIZE=integer. It crashed at
  memcpy(&lf_data->planes[i].dst, &pd[i].dst,
         sizeof(lf_data->planes[i].dst));
That is in loop_filter_data_reset() function.
It seemed that MOVAPS was used in memcpy(), which required that the
memory operand must be aligned on a 16-byte boundary.

First fix(patch set 1): align lf_sync->lfdata to 16 bytes during
memory allocation.

Change-Id: I2d2d4bae1c1546ed7b7e61bafea2334124079719