Fix handling of annexb + S2T1 format streams
Rework the main loop in decoder_decode() so that, in annexb format,
only one frame's worth of data at time is passed to decode_one().
This fixes an issue when decoding S2T1 format streams.

It appears that decoder_decode() had some code left over from
the old VP9 superframe mechanism. This is no longer useful, so
remove that code while we're here.

BUG=aomedia:1872

Change-Id: I10a322bb034cabdeadbe393d9c73217a738ee3db