Check frame_refs_short_signaling before written
The encoder may turn on frame_refs_short_signaling if needed.
Currently we only allow it on when all the reference frames
are distinct and there are indeed 7 reference frames available.
If not, the frame_refs_short_signaling flag will be turned off.
Hence, the flag of frame_refs_short_signaling should be written
to the bitsteram with its correct value after the above checking.

BUG=aomedia:1816

Change-Id: Ia3c0acb8c656cc27318bfcc4de83e962c1f4ee04