Superblock size choosing in large scale tile
The superblock is a sequence parameter and has to be the same for the
whole sequence. In lightfield application, we must choose the super-
block size(either 64x64 or 128x128) before the encoding starts.
Otherwise, the default is AOM_SUPERBLOCK_SIZE_DYNAMIC, so the encoder
will decide the superblock size.

Change-Id: I210e1c978e1706dabe5a68580c5320ab775f6957