Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

I don't know if the tier is meaningful for still pictures or not?

Are the level constraints on bitrate meant to impose any restriction on a still picture?
(My guess is that they do not apply and so there is no need to send the tier - i.e. no changes needed to your patch, but I just wanted to check this was an intentional omission.)

I also have mixed feelings about not sending the tier for low levels as it makes it a little harder to adjust the levels and order of operating points.  On the other hand, the tier performance levels are not defined for low levels so it also seems pointless to send a bit we will never use.

Patch-set: 2
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+1