Update patch set 2
Patch Set 2:

This looks fine, but it made me wonder if it's legal to have a padding OBU that's unaccounted for by the TU and frame sizes in an annex-B stream. For example, in the case where:

- There's only one frame in the 
- TU data_sz is greater than TU size
- frame_size is less than TU size

Do we have a valid stream if there's a padding OBU, or all zeros written in that remaining space (the space at the end of data, size of data_sz - TU size)?

Patch-set: 2
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>