cmake: Set CMAKE_AR to ${CROSS}ar for mingw
Likewise for CMAKE_RANLIB and ${CROSS}ranlib.

The extra "CACHE FILEPATH <helptext>" has to be set, otherwise
cmake treats it as an absolute path (which isn't found) and
CMAKE_AR remains set to an empty string instead.

This makes sure the right cross tools are used, instead of using
the system default 'ar' tool in /usr/bin/ar.

This fixes cross compilation for mingw from macOS.

Change-Id: Ie8729c2209358cb58b8e457df562f3908b5f4747