Update patch set 2
Patch Set 2:

The reason for not sending the tier for lower levels was to make sure that no high tier is signaled for them, and to save some bits too. However, if it seems preferable to always signal the tier for alignment, I would be fine with that too.

Patch-set: 2
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>