Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> Why adding this new restriction ? Clamping should be enough or am i missing something?

Fred: it was agreed at the last WG call to change the last sentence in section 7.12.3 "Reconstruct process" of the spec into an informative note. The sentence currently reads "It is a requirement of bitstream conformance that the values written into the Residual array in step 2 are representable by a signed integer with 8 + BitDepth bits." The question is what the maximum range of the reconstructed residuals is given all the existing restrictions on the bitstream and clamping operations. The intent here should be to summarize existing restrictions, not create a new one.

Patch-set: 4
Reviewer: Gerrit User 8487 <8487@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>