Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> Hi Tom,
> 
> Both of your suggestions, especially the first one, are fine by me. It is uncommon to see source files at the top level of a source tree, except for files like aom_config.h that truly apply to the entire source tree.
> 
> The current workaround of including "./aom_config.h" is very confusing. It can be misinterpreted to mean aom_config.h is in the current directory.
> 
> By the way, I found aom_ports/config.h, which includes "aom_config.h" (just the file name).

If this patch is something we decide to land I'll deal with aom_config.h in a follow up. That one is going to be a huge amount of noise because it's included in a ton of places (all via ./aom_config.h). I'll change CMake to put aom_config.h in config/aom_config.h, or perhaps generated_config/aom_config.h, and then update all of the include sites.

I'm not sure of the origins of aom_ports/config.h, but it can likely be removed from the repository (as part of the aom_config phase of this clean up).

Patch-set: 4
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>