Initial encoder implementation for fwd kf
Add encoder implementation to code forward reference
keyframes as a unfiltered altrefs. In this case, the altref source
is placed on the keyframe location determined in firstpass.c.
This frame is first coded as an invisible keyframe which only updates
the altref buffer. All reference buffers are cleared when the keyframe
is finally displayed using cm->show_existing_frame.

This is a preliminary implementation and has not yet made appropriate
changes to the rate control. There also seems to be a bug in some
instances when the keyframes are spread far apart, and seems to be
an issue with the way the postencode_update is being handled. I am
currently working on a fix as a followup.

This patch does not impact results if --enable-fwd-kf is set to 0.

Preliminary Results:
20 encoded frames with kf-max-dist set to 5: -9.47% gain
20 encoded frames with kf-max-dist set to 8: -11.02% gain
20 encoded frames with kf-max-dist set to 10: 1.60%  loss

STATS_CHANGED

Change-Id: I598038162756a473a033e8d1c26715b0429844a3