Fix infinite loop in --all-layers mode
When in --all-layers mode, obudec.c "splits" temporal units into
individual spatial layers. Due to the way this works in non-annexb
streams, it is possible for a split to be placed immediately after
a (redundant) sequence header OBU.

This means that, eventually, aom_decode_frame_from_obus() gets passed
a buffer containing a single sequence header OBU and nothing else.
Unfortunately, if this happens, it does not update *p_data_end properly,
and the result is that decoder_decode() enters an infinite loop.

This patch fixes that unintended behaviour.

BUG=aomedia:1923

Change-Id: I0860de4925d76b4126a9f3bbf5677ccc2e7b30c4