Remove the raw_data_end field from TileBufferDec.
The caller of get_ls_tile_buffers() was only using the raw_data_end
field of the last tile buffer
(tile_buffers[tile_rows - 1][tile_cols - 1].raw_data_end). Change
get_ls_tile_buffers() to return that value as the function return value.

This shaves off 1/3 of the size of the tile_buffers array in AV1Decoder,
a saving of 32 KB in 64-bit build, or 16 KB in 32-bit build.

BUG=aomedia:1870

Change-Id: I9ebfd396372df0e56ec823a3426353c5993185be