[Normative] Update xd->qindex and xd->lossless at SB level
In current codebase, xd->qindex[] is only set up once at frame
level for all the segments. This qindex is used to determine
whether we need to decode the normal transform type or just use
lossless transform. However, if deltaq and aq segment are both on,
the actual qindex is determined based on the modulated q of the
superblock. In other words, mbs assigned the same segment id might
use different qindex because sb-level q can be different.

To fix it, we update xd->qindex[] array at superblock level once
delta-q of a superblock has been determined. Similar update is
done for xd->lossless[].

BUG=aomedia:1925

Change-Id: Ife1373a2250e6aa214a4385088eba3f6d3aadd6d