Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Out of curiousity: is there any reason why the neon assembly would be useful if hooked up?
> Or we can't because of some reason like (i) they are not matching with C versions and/or (ii) we would rather have intrinsic versions of these?

I'm not sure why/when they were originally disabled but I suspect it was because the C versions were changed or removed.

In general we prefer intrinsics because armv7 and arm64 assembly syntax differs and we don't want to maintain two versions of the assembly. We have deleted all the assembly in libvpx except a few corner cares where the assembly is faster on armv7. Those tend to be very small simple functions which operate on 4x4 to 8x8 blocks. The intrinsics for loads/stores can be difficult to manage in those situations.

In every other case the intrinsics matched the assembly or outperformed it. It's also much easier to read, write, test, and modify.

Patch-set: 2
Reviewer: Gerrit User 5435 <5435@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>