decoder: Rework cm->frame_to_show usage
Some upcoming changes (see the attached issue) will cause the encoder
and decoder to have different ways of outputting frames. Thus it
no longer makes sense to have a common "frame_to_show" field.

Thie commit moves the decoder logic over to using a new field
pbi->output_frame instead of cm->frame_to_show. In addition, it
reworks the life cycle of frames inside the decoder.

The main difference is that whatever is stored in pbi->output_frame
persists until the next time that field is written to, rather than
always being freed at the start of the next frame. This does not
make much difference for the moment, but will be important when
the decoder needs to queue up multiple output frames.

BUG=aomedia:1941

Change-Id: I38289c3fd051d2692cde397146ae77e577a45ba8