Remove 'last_layer_id' from ObuDecInputContext
Previously, if we wanted to use all-layers mode, the code
in obudec.c was responsible for carving up temporal units
into spatial layers to feed into the decoder. Whether we were
in this mode corresponded to whether last_layer_id is set to
IGNORE_ENHANCEMENT_LAYERS or not.

Now, the all-layers behaviour is implemented in the decoder proper,
so we can always pass a full temporal unit at a time into the decoder.

This means that the special code in obudec.c is no longer needed,
and that all-layers mode now works correctly for "contained" streams
(ie, annexb format, .ivf, .webm, etc.)

BUG=aomedia:1941

Change-Id: I38a21e560c14f822b652f8c2880f6be9f7c2e2ec