variance based aq modification
This changes aq from using an actual block variance to using
the log of a local 4x4 variance to avoid situations in which
a large block with a gentle gradient gets marked high variance
even though each subblock has a low variance. This change enables
us to assign the same segment number for the same sorts
of area regardless of the partitioning.

This change required tweaking some parameters for where each
level lands.

Also - uses rdmult for costing when coming back to higher level block
so that the partition decision is based on rdmult from high level block
skirting the issue that each subblock could be using a different
rdmult.

Change-Id: Ie8aa00ef20eb41dd15c2c1f62ec2b29312a9bbe7