Allocate space for the largest types in intrinsics tests
This stops the ARM gcc 7.3.0 warnings about the use of indices out of
bounds.  Only the space needed for the types was allocated and correct
access to that space is ensured by if tests for the types, but some
gcc versions are doing the out of bounds checks before unreachable
code is eliminated, and will therefore warn.

Asserts have been added to ensure that enough space does get
allocated.

BUG=aomedia:1313

Change-Id: I5f64f71f88ce2e0f26b316919b930861c24108c6