Update patch set 3
Patch Set 3:

Hi Iole,

Yes. This was normative changes. At first, I thought I was matching what is defined in the spec, and no new proposal was involved, "normative" was missing for the first 3 patches. Sorry for that.

As of today, the total patches regarding the ext-tile implementation that have been merged are 19, and 9 of them are normative ones. Here is the list:
(At first, I thought I was matching what is defined in the spec, and no new proposal was involved, "normative" was missing for the first 3 patches.)
[NORMATIVE]ext_tile: setting single_tile_decoding flag
[NORMATIVE]ext_tile: remove a large_scale_tile dependency
[NORMATIVE]ext_tile: set large_scale_tile coding flag
[NORMATIVE] ext_tile: write/read tile information
[NORMATIVE] ext_tile: set external references at the tile level
[NORMATIVE] ext_tile: decode the tile list OBU
ext_tile: decode the common camera frame header
ext_tile: handle the skipping of the frame data
Add OBU_TILE_LIST obu type

Thanks,
Yunqing

Patch-set: 3