[NORMATIVE] Always check that mvs are in range
Move the check of is_mv_valid() so that it's always called at the
end of assign_mv(), instead of only calling it for NEWMV type modes.
This will allow catching any cases where invalid motion vectors
are generated.

This is normative in that it disallows streams which generate "invalid"
motion vectors via a non-NEWMV mechanism (eg, GLOBALMV). But that was
never intended to be allowed anyway.

BUG=aomedia:1959

Change-Id: I77ad1b841a84b1b3de852b7ef0ba42713554c7a6