ext_tile: handle ext-tile tile info better
This patch includes the following changes.
1. Store the ext-tile tile information as a separate byte at the end
of the frame header. That information is only needed for parsing the
ext-tile bitstream. While constructing the camera frame header during
the bitstream parsing, this byte is dropped. Therefore, the final AV1
decodable bitstream doesn't include ext-tile tile information in the
frame header. This change only affects ext-tile bitstream, so it is
non-normative.
2. Add an API to get frame header information.
3. Improved the bitstream parsing example to only copy the camera
frame header to the final bitstream.

Change-Id: I7db43dfe05360e874c002725b3ed8af90143c8b2