cmake:Add option ENABLE_TESTDATA
Add option to config if CMake should generate
targets to download testdata. Default value is ON.

There are many testdata targets, while debugging
with some IDE, these targets are very annoying.
This option make it possible to turn them off when
they are not needed. But by default, nothing changed.

Change-Id: I97850e28eef59b7bd4ed93a65d553df481e636a4