Adds a DNN regression based rate distortion model
Implements a deep neural network based regressor for rate and
distortion meant to be agnostic of transform partitions and
transform types for use in pruning and early breakouts in
various stages of the rd search at the prediction unit level.

The DNN regressor uses 20 easy to compute features with 2 hidden
layers of 10 nodes each, for both rate and distortion.

The RMSE for both rate and distortion achieved on a validation set
is about a quarter of that achieved with the old model.

The new model is not yet hooked up in the rd decision mechanism,
so this patch does not change the coding results.

Change-Id: I33d5ea08f6125e7b36f74fca5b89d44386078805