Adds a DNN regression based rate distortion model
The DNN regressor uses 20 easy to compute features with 2 hidden
layers of 10 nodes each for each of rate and distortion.

The RMSE for both rate and distortion is about a quarter of that
achieved with the old model.

The new model is not yet hooked up in the rd decision mechanism,
so tnhis patch does not change the coding results.

Change-Id: I33d5ea08f6125e7b36f74fca5b89d44386078805