[NORMATIVE] Fix random-access decoding
To conform the the specification, we need to be able to decode
sequences beginning with a forward keyframe, followed by a show-existing
frame displaying that keyframe.

However, we currently can't do that if frame IDs are in use. The code
currently assumes that a previous frame ID is available for every frame
that is not a shown keyframe. So for the initial forward keyframe,
we think that there is a previous frame with ID 0. This causes things
to break if the forward keyframe's ID is too far away from 0.

The fix is to explicitly track whether we've decoded any frames yet,
and to allow the first frame decoded to use any frame ID value.

BUG=aomedia:1968

Change-Id: I6346dd070d7b3615c24bef63b332d8901bf13aef