Update patch set 2
Patch Set 2:

Hi Deepa and Ranjit,
To avoid the duplication in encoding in the tests, can we encode once and store the result? Or, even pre-encode the file and then upload it as the test vector for MT decoder test?

Patch-set: 2
Reviewer: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>