Refactor: Clear the logic of buffer updates
The buffer update configurations set in the function
av1_get_compressed_data are only used in the firstpass (as
in the second pass they  will be over-written by the function
configure_buffer_update).

This patch explictly specifies that those configurations are
used only in the first pass, and also cleans up an unused variable.

Change-Id: I565ed0f1db1875cb38cc08242a94ec8e546575a8