Update patch set 2
Patch Set 2:

I believe:
1) for valid streams this change has no effect
2) for invalid streams the code will now silently ignore some errors that used to be caught before the change

Because of point 1, I don't think this change needs to be marked as Normative.  (Unless you have some valid stream that demonstrates a problem?  If so, please attach/email it to me so I can check it against the spec.)

Because of point 2, I am not sure whether this change is needed at all.
On the one side, it is nice to be robust to a greater range of invalid streams, on the other side, it is nice for the reference decoder to generate warnings/errors when an invalid stream has been constructed.

Patch-set: 2
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>