enc_worker_hook should return 1 on success.
enc_worker_hook should return 1, not 0, on success. Since
av1_encode_tiles_mt() ignores the return value of enc_worker_hook()
(because it ignores the return value of sync()), the return value
doesn't really matter. But it is better to return the correct value.

Clarify the comment that documents the return value of AVxWorkerHook.

Clean up the code in av1_encode_tiles_mt() that sets up worker->data1
for enc_worker_hook().

Clarify the comment that documents the return value of AVxWorkerHook.

Change-Id: I635141fb2b1df05ef56a62bec991e0cac970cd3c