Remove the uncomp_hdr_size field of AV1Decoder.
The uncomp_hdr_size field of AV1Decoder is only used by
av1_decode_frame_headers_and_setup to return the frame header size to
the caller. It is better to use an output parameter or the return value
to do that.

Also document the return values of read_frame_header_obu and related
functions.

Change-Id: Iee5f5f5763e076cedf9316d532a6e49f412fca97