Change reference buffer update rules when new gf struture used
In the new gf structure, backward reference frames are always coded
in reverse order (i.e. farthest frames will be coded first). Whenever
a backward reference is coded, we first shift the reference frames
as follows:

BWDREF -> ALTREF2-> EXT_REF,

and then have BWDREF refreshed by the newly coded frame. When an overlay
frame to be shown, we display the frame stored in BWDREF and shift the
reference as follow:

BWDREF <- ALT2REF <- EXT_REF,

so that BWDREF points the next backward reference frame to be shown.

Change-Id: Ie615fb4e7c1e4f1a3440468b2e6525f3540d9783