dc_predictor_rect: Assert for valid input only.
This assert is mainly to ensure that divide_using_multiply_shift does
not need clamping, if the input arrays 'above' and 'left' have values in
valid range.

In certain unexpected use of public APIs, it's possible for inputs to be
invalid. So, we shouldn't check the assert in those cases.

BUG=oss-fuzz:9323,9327

Change-Id: I3e06bc8d2624189d37183aa0c879b5521b524683