Put IntraPredFunc outside the anonymous namespace.
This works around a strange compiler warning in the begin() and end()
methods of the ParamGenerator template class in
gtest/internal/gtest-param-util.h:
  warning: ‘<anonymous>’ is used uninitialized in this function

Another workaround is to put the entire file in a named namespace
(such as "namespace aom_internal") instead of the anonymous namespace.

BUG=aomedia:2003

Aside of cherry-picking, this patch fixes cmake-format issues for
all cmake files.

Change-Id: I32c31ee1ebd51495914b9ccaf5c01dbe9d9a2df5