Update patch set 1
Patch Set 1:

There is a clamp for those values, when generating grain values so the error should not appear again. The problem is that when the grain was not going to be used in the future, it was not initialized to zero (or to anything). I added setting it to 0 in another patch

Patch-set: 1
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>