Document how to call setjmp() correctly.
The jmp_buf after a setjmp() call is valid only for the duration of the
function that calls setjmp(). This is documented in:
http://pubs.opengroup.org/onlinepubs/9699919799/functions/longjmp.html
https://en.cppreference.com/w/c/program/longjmp

Therefore the function that calls setjmp() must reset the 'setjmp'
boolean field to 0 (to indicate that the jmp_buf is invalid) before it
returns.

The call_setjmp() function removed in commit
f3ac65169d3ab435725aff888045617bc6c8ca8e did not meet this requirement
because it set the 'setjmp' field to 1 before one return statement,
therefore it was incorrect.

BUG=oss-fuzz:9463

Change-Id: I2f8a7e73910d298a961ebd3c9c1c371fa0ffe121