Update patch set 1
Patch Set 1:

> I agree that this matches the spec, but I don't know why we have the rule since it disallows usage of OBU_FRAME for a keyframe dependent recovery point. Anyone able to shed a little light on that?
> 

OBU_FRAME indicates the OBU contains a frame header and frame data.
show_existing_frame equal to 1 indicates the OBU will not have any frame data.

There was a concern that it was unclear whether frame data is present or not for the case when both of these are signalled (i.e. OBU_FRAME && show_existing_frame==1).  (In particular it is unclear whether trailing bits have to be added.  They are not normally added for OBU_FRAME as they are present in the frame data, but would need to be added for this special case.)

Therefore to avoid this confusion it was recommended to make the ambiguous combination illegal.

Patch-set: 1
Reviewer: Gerrit User 5260 <5260@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>