Fix the quality boost ratio between ALT2 and ALT in AOM_Q mode
Originally, the best q index of each frame is computed based on
the q index used in the previous frame. Therefore, it is possible
that the quality boost for the second level ARFs can be higher than
the quality boost for ALT0.

This patch uses the q index for the ALT0 as the baseline and decrease
the quality boost for the second level ARFs by half.

Coding gain in overall psnr (speed 1):
         17 frames    97 frames
lowres:    0.064        0.047

STATS_CHANGED

Change-Id: I2faba9a53526df1e9e4ed90c1644556009c3f983