ext-tile: output decoded tiles in output frame
As defined in the Spec, the decoded tiles in a tile list are stored in
the output frame buffer.

The variable outputW is defined as
  ( 1 + output_frame_width_in_tiles_minus_1 ) * TileWidth.
The variable outputH is defined as
  ( 1 + output_frame_height_in_tiles_minus_1 ) * TileHeight.
The operation of writing a decoded tile into the output frame in raster
order.

This patch re-arranged output data as defined above.

BUG=aomedia:2047

Change-Id: I8a0deeb89e387c13573ca7cb5a0b4125d10989d9