Update patch set 1
Patch Set 1:

Hi Yungqing,

Thanks for this.  On the face of it, this looks good and I'm happy to test with it, but will need two changes first:
1). Please see https://bugs.chromium.org/p/aomedia/issues/detail?id=2051  You'll need to make this change in the aom_shift_img() function as well.  Maybe worth basing your patch on top of that one?
2). Would you implement for lightfield_tile_list_decoder.c too please? - that's the one I've got streams to test with.

Is your intention to get this pushed to av1-normative too?  I'd like the behaviour to be the same on both branches.

Kind regards,
Jonathan

Patch-set: 1
Reviewer: Gerrit User 6145 <6145@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>