Update patch set 2
Patch Set 2:

Hi Yunqing,

Thanks for this patch.  I won't have time to look in detail tonight, but just a quick question.  Does this patch meet our request that I put on the bug report?
"Our preference would be that:
1). the library follows the spec (i.e. fills the decoded tiles into the output frame)
2). the default output from lightfield_tile_list_decoder remains the same (i.e. the 1D tile data, which after #1 will be sampled from the output frame) - we are using lightfield_tile_list_decoder for conformance testing, so changes to the output yuv file are painful!

I haven't cross referenced with the specification, but code I saw that zeroes the output buffer in obu.c doesn't seem to correspond with the following statement in the specification:
"The output frame may not be fully covered with decoded tiles. The decoder should not modify samples in the output frame outside of the boundaries of the decoded tiles."

Given that the decoder doesn't specify how the anchor frames are provided, the contents of the output frame prior to the first tile list obu being specified, and hence the contents of the output frame when not fully covered by decoded tiles, are undefined by the specification.  As such, it doesn't make sense to output the entire output frame (request #2 above).

Kind regards,
Jonathan

Patch-set: 2
Reviewer: Gerrit User 6145 <6145@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>