Bitmask patch 6: Minor improvement.
(1). Update bitmask to the exact bit, no need of for loop.
(2). Remove operations to point to odd chroma plane position.
Note that if there's subsampling, chroma plane position corresponds
to the odd position of luma plane. But in our case, we build the
bitmask, instead of setting the bit of chroma odd position, we set
it at the bit of even position. This will not change the behavior
of code, but save some operations and thus faster.

Change-Id: I8536e2b0511277df7944920d9e598edb60cebdef