Integerized Wiener encoder (fixed saturate mul)
DO NOT MERGE

I'm submitting this to test against AWCY, these changes are not ready to merge.

This change includes the higher precision tap storage and uses saturating
arithmetic in a few strategic places, resulting in apparent improvement.  Here
I fixed the saturating multiply

I will be developing an integer-only implementation of the Wiener
reconstruction filter encoder, partly as a useful reference for hardware
implementations, partly to investigate possible speed-up.

* At the moment the code is a mess due to changes being wrapped in a build
  flag. Assuming there's no compression performance regression I'll clean up
  later by removing the build flag and replacing the originals.
* Integerized version currently only supports 8 bit depth, so I disable highbd
* SIMD vectorized functions aren't integerised yet, so I bypass them

Change-Id: Ia5a5033e39ff134d15521484fb82ca097b48eb07