Implement aom_reader_has_error().
Although the aom_reader_has_error() function already exists, it wasn't
actually implemented and would always return 0 (false), because the
underlying dec->error field was initialized to 0 and never changed.

The only error we report is the decoding overflow error. This makes the
the aom_reader_has_overflowed() function unnecessary, so that function
is removed.

BUG=oss-fuzz:9663

Change-Id: Ia77141754c158e70a92e7770ceb38ea2ae92c153