Pass n to get_msb in the primitive_quniform funcs.
The functions this CL modifies were added in the following two CLs:
https://aomedia-review.googlesource.com/c/aom/+/8789
https://aomedia-review.googlesource.com/c/aom/+/20820

These functions pass n - 1 to get_msb(). I believe these functions
implement the ns(n) and NS(n) descritors of the AV1 spec. ns(n) and
NS(n) pass n, not n - 1, to FloorLog2(). (FloorLog2() is the same as
libaom's get_msb() function.)

A careful analysis of the code shows passing n instead of n - 1 to
get_msb() results in the same encoding. Passing n matches the AV1 spec
and is slightly more efficient when n is a power of 2.

Change-Id: Ibeff837059e9c9de2b4ab9cb857a2ea60b1c4767