Add a macro to reduce border extension.
Can be used to reduce encoder memory consumption, when superres and
resize are not being used.

Overall encoder memory reduction:

(i) With CONFIG_LOWBITDEPTH=0
-  480p video: 11.2%
- 1080p video:  8.2%

(ii) With CONFIG_LOWBITDEPTH=1
-  480p video: 7.4%
- 1080p video: 5%

Note: This is a short-term solution. Later, with better testing in
place, we can try to pick the appropriate border extension at runtime.

BUG=aomedia:2108

Change-Id: Icf4a8609692fd630d397c2dd4b910b01cafa58f5