Bound the total allocated memory of frame buffers.
By setting a debugger breakpoint in aom_memalign() (called by both
aom_malloc() and aom_calloc()) for 'size' greater than 10 MB, I found
that the frame buffer pool is one of the two biggest consumers of heap
memory in the decoder. This CL tries to bound the memory consumption of
the frame buffer pool. We could define a new macro for the maximum frame
buffer pool size, but for now I am borrowing the existing
AOM_MAX_ALLOCABLE_MEMORY macro for this purpose.

This CL only does an approximate memory allocation size check and
ignores the following complications:
- The frame buffer pool may have as many as FRAME_BUFFERS
  (= REF_FRAMES + 7), not REF_FRAMES, frame buffers.
- If use_highbitdepth is true, aom_realloc_frame_buffer() also allocates
  a yplane_size-byte buffer as ybf->y_buffer_8bit.

BUG=oss-fuzz:9956
BUG=oss-fuzz:9992
BUG=oss-fuzz:9993

Change-Id: Iab894920bb5e85aa13d92598fd18e88c84ae64fd